+++ lib/main.dart
@@ -1,13 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
-import 'package:flutter_markdown/flutter_markdown.dart';
-import 'package:google_generative_ai/google_generative_ai.dart';
 import 'package:provider/provider.dart';
+import 'package:shokhi/pages/navigation_screen.dart';
 import 'package:shokhi/provider/theme_provider.dart';
 
 void main() async {
   await dotenv.load(fileName: ".env");
+
   runApp(
     MultiProvider(
       providers: [
@@ -44,329 +43,7 @@ class MyApp extends StatelessWidget {
         ),
         useMaterial3: true,
       ),
-      home: const Home(),
-    );
-  }
-}
-
-class Home extends StatefulWidget {
-  const Home({super.key});
-
-  @override
-  State<Home> createState() => _HomeState();
-}
-
-class _HomeState extends State<Home> {
-  @override
-  Widget build(BuildContext context) {
-    bool isDarkMode =
-        Theme.of(context).colorScheme.brightness == Brightness.dark;
-
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(
-          "ಠ﹏ಠ Shokhi ಠ﹏ಠ",
-          style: TextStyle(
-            color: isDarkMode ? Colors.white : Colors.black,
-          ),
-        ),
-        centerTitle: true,
-        backgroundColor: Colors.transparent, // Transparent background
-        elevation: 20.0, // Remove shadow
-        iconTheme: IconThemeData(
-          color: isDarkMode ? Colors.white : Colors.black,
-        ),
-        leading: Builder(
-          builder: (context) => IconButton(
-            onPressed: () => Scaffold.of(context).openDrawer(),
-            icon: Icon(
-              Icons.menu,
-              color: isDarkMode ? Colors.white : Colors.black,
-            ),
-          ),
-        ),
-        actions: [
-          IconButton(
-            onPressed: () {
-              context.read<ThemeProvider>().toggleTheme(context);
-            },
-            icon: isDarkMode
-                ? const Icon(Icons.dark_mode_rounded)
-                : const Icon(Icons.light_mode_rounded),
-          ),
-          // A little bit of spacing to the right
-          const SizedBox(width: 10),
-        ],
-      ),
-      drawer: Drawer(
-        child: SafeArea(
-          child: Column(
-            children: [
-              const UserAccountsDrawerHeader(
-                currentAccountPicture: CircleAvatar(
-                  backgroundColor: Colors.grey,
-                  child: Text("S"), // Replace with user image or initials
-                ),
-                accountName: Text("Shokhi User"),
-                accountEmail: Text("shokhiuser@example.com"),
-              ),
-              ListTile(
-                leading: const Icon(Icons.login),
-                title: const Text("Login with Google"),
-                onTap: () async {
-                  // Implement Google Sign-in logic here
-                  // You'll need to install and use the `google_sign_in` package
-                  // Refer to https://pub.dev/packages/google_sign_in for details
-                },
-              ),
-            ],
-          ),
-        ),
-      ),
-      body: const ChatScreen(),
-    );
-  }
-}
-
-class ChatScreen extends StatefulWidget {
-  const ChatScreen({super.key});
-
-  @override
-  State<ChatScreen> createState() => _ChatScreenState();
-}
-
-class _ChatScreenState extends State<ChatScreen> {
-  late final GenerativeModel _model;
-  late final ChatSession _chat;
-  final TextEditingController _textController = TextEditingController();
-  final ScrollController _scrollController = ScrollController();
-  bool _loading = false;
-
-  @override
-  void initState() {
-    _model = GenerativeModel(
-      model: "gemini-pro",
-      apiKey: dotenv.env['API_KEY']!,
-    );
-    _chat = _model.startChat();
-    super.initState();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    bool hasApiKey =
-        dotenv.env['API_KEY'] != null && dotenv.env['API_KEY']!.isNotEmpty;
-
-    return Padding(
-      padding: const EdgeInsets.all(8.0),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Expanded(
-            child: hasApiKey
-                ? ListView.builder(
-                    controller: _scrollController,
-                    itemBuilder: (context, idx) {
-                      final content = _chat.history.toList()[idx];
-                      final text = content.parts
-                          .whereType<TextPart>()
-                          .map<String>((e) => e.text)
-                          .join('');
-                      return MessageWidget(
-                        text: text,
-                        isFromUser: content.role == 'user',
-                      );
-                    },
-                    itemCount: _chat.history.length,
-                  )
-                : ListView(
-                    children: const [
-                      Text('No API key found. Please provide an API Key.'),
-                    ],
-                  ),
-          ),
-          Padding(
-            padding: const EdgeInsets.symmetric(
-              vertical: 9,
-              horizontal: 9,
-            ),
-            child: Row(
-              children: [
-                Expanded(
-                  child: TextFormField(
-                    controller: _textController,
-                    autofocus: true,
-                    decoration: InputDecoration(
-                      contentPadding: const EdgeInsets.all(11),
-                      hintText: 'Ask shokhi anything...',
-                      border: OutlineInputBorder(
-                        borderRadius: const BorderRadius.all(
-                          Radius.circular(14),
-                        ),
-                        borderSide: BorderSide(
-                          color: Theme.of(context).colorScheme.secondary,
-                        ),
-                      ),
-                      focusedBorder: OutlineInputBorder(
-                        borderRadius: const BorderRadius.all(
-                          Radius.circular(14),
-                        ),
-                        borderSide: BorderSide(
-                          color: Theme.of(context).colorScheme.secondary,
-                        ),
-                      ),
-                    ),
-                    onFieldSubmitted: (String value) {
-                      _sendChatMessage(value);
-                    },
-                  ),
-                ),
-                const SizedBox.square(
-                  dimension: 15,
-                ),
-                if (!_loading)
-                  IconButton(
-                    onPressed: () async {
-                      _sendChatMessage(_textController.text);
-                    },
-                    icon: Icon(
-                      Icons.send,
-                      color: Theme.of(context).colorScheme.primary,
-                    ),
-                  )
-                else
-                  const CustomLoadingAnimation(),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Future<void> _sendChatMessage(String message) async {
-    setState(() => _loading = true);
-
-    try {
-      final response = await _chat.sendMessage(Content.text(message));
-      final text = response.text;
-      if (text == null) {
-        debugPrint('No response from API.');
-        return;
-      }
-      setState(() => _loading = false);
-    } catch (e) {
-      debugPrint(e.toString());
-    } finally {
-      _textController.clear();
-      setState(() => _loading = false);
-    }
-  }
-}
-
-class CustomLoadingAnimation extends StatefulWidget {
-  const CustomLoadingAnimation({super.key});
-
-  @override
-  State<CustomLoadingAnimation> createState() => _CustomLoadingAnimationState();
-}
-
-class _CustomLoadingAnimationState extends State<CustomLoadingAnimation>
-    with SingleTickerProviderStateMixin {
-  late AnimationController _controller;
-  late Animation<double> _scaleAnimation;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = AnimationController(
-        vsync: this, duration: const Duration(milliseconds: 500))
-      ..repeat();
-    _scaleAnimation = Tween<double>(begin: 1.5, end: 0.5).animate(_controller)
-      ..addListener(() => setState(() {}));
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Icon(Icons.send, size: 24 * _scaleAnimation.value);
-  }
-}
-
-class MessageWidget extends StatelessWidget {
-  final String text;
-  final bool isFromUser;
-
-  const MessageWidget({
-    super.key,
-    required this.text,
-    required this.isFromUser,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment:
-          isFromUser ? MainAxisAlignment.end : MainAxisAlignment.start,
-      children: [
-        // Use the custom TextContainerWithCopy widget
-        TextContainerWithCopy(isFromUser: isFromUser, text: text),
-      ],
-    );
-  }
-}
-
-class TextContainerWithCopy extends StatelessWidget {
-  final bool isFromUser;
-  final String text;
-
-  const TextContainerWithCopy(
-      {super.key, required this.isFromUser, required this.text});
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisSize: MainAxisSize.min, // Ensures content fits
-      children: [
-        if (isFromUser)
-          IconButton(
-            onPressed: () => Clipboard.setData(ClipboardData(text: text)),
-            icon: const Icon(Icons.content_copy),
-            iconSize: 18.0,
-            color: Colors.grey, // Adjust color as needed
-          ),
-        Container(
-          constraints: const BoxConstraints(maxWidth: 320),
-          decoration: BoxDecoration(
-            color: isFromUser
-                ? Theme.of(context).colorScheme.primaryContainer
-                : Theme.of(context).colorScheme.surfaceVariant,
-            borderRadius: BorderRadius.circular(18),
-          ),
-          padding: const EdgeInsets.symmetric(
-            vertical: 15,
-            horizontal: 20,
-          ),
-          margin: const EdgeInsets.only(bottom: 8),
-          child: MarkdownBody(
-            selectable: true,
-            data: text,
-          ),
-        ),
-        if (!isFromUser)
-          IconButton(
-            onPressed: () => Clipboard.setData(ClipboardData(text: text)),
-            icon: const Icon(Icons.content_copy),
-            iconSize: 18.0,
-            color: Colors.grey, // Adjust color as needed
-          ),
-      ],
+      home: const NavigationScreen(),
     );
   }
 }
